Charlie Weapons Squad Kit Update
Removed
Zip ties from assistants and gunners (to stop the accidental arrest of the gunner)
All the different smoke colors that are not in the BN pace plan.
Removed the tinted glasses for all the kits.
Removed spray can from assistants
Stinger assistant kit
240 Ammo Bearer kit (240 Assistant kit also covers this kit)
MAAWS Ammo Bearer Kit (MAAWS Assistant also cover this kit)

Added:
Tripod and spare barrel for 240 assistant
Headsets for all the helmets
Added Red and Blue smoke to FTLs and WSLs to be standardized with the BN pace plan.
Added an extra HE round to the MAAWS assistant so he now carries 2 HE and 2 HEAT and is still on the weight limit.